refactor(gh-report): rewrite domain usize/isize to u64/i64 per GEN-0004:R1 (commit-4 package step 1 of 3)
Sub-4a of mission f2d-writer-che-nnnn-1747526400. Pure type-shape change at
DomainEvent variant fields, Run/Repo aggregate fields, and command structs;
constructor-site conversions inserted at producer boundary. No behavioural
change. Tidy-First; sub-4b adds Encode impls, sub-4c runs workspace verify.

Oracle bead adr-fmt-3e9o (corpus unanimous): GEN-0004:R1 forbids usize/isize
in serialized event-field type positions (platform-dependent width breaks
determinism); GEN-0032 + GEN-0037 EVT-004/EVT-005 require append-only
event-shape evolution; PAR-0011 64-bit gate makes usize→u64 lossless on every
pardosa target; COM-0023 governs conversion-site discipline.

Conversion sites (9 total — full table in review bead adr-fmt-4fk7):

  app/collect.rs:544,550,713  Vec::len() → u64       as u64 (PAR-0011 R1)
  app/collect.rs:673          StartSweep.repo_count  as u64 + comment (R1)
  app/collect.rs:829          CompleteSweep          as u64 (R1)
  app/collect.rs:843          PublishEvidence.pages  as u64 + comment (R1)
  app/collect.rs:911          completed (Proj::len)  as u64 + comment (R1)
  app/collect.rs:1402         warm_repo_count (u32)  u64::from (R2 widen)
  app/collect.rs:1541         PublishEvidence.pages  as u64 (R1)
  app/collect.rs:1728         pending_repos          let : u64 = 0
  app/collect.rs:1745         RenderPartial.pages    as u64 (R1)

Pattern: u64::from(_) for infallible widening from narrower fixed-width
integers (COM-0023:R2 happy path); as u64 from Vec::len()/collection sizes
with PAR-0011 64-bit-gate justification. No try_from needed — no untrusted
external sources in this diff.

Repo::apply evaluation_count: += 1 → saturating_add(1). Justification per
CHE-0009:R1 (apply must not panic): checked_add returning None gives no
replay recovery (the event is already in the log); wrapping_add silently
corrupts state; saturation at u64::MAX (~1.8e19) is the only correct choice
under the infallible-apply contract, and the saturation point is practically
unreachable.

Tidy-First chain (prior commits in mission f2d-writer-che-nnnn-1747526400):

  6aefe96  CHE-0064 ratify (supertrait gate)
  468e1f4  cherry-pit-core Encode (commit-2a, adr-fmt-hawc)
  1d2cea3  5-crate Encode propagation (commit-2b, adr-fmt-dtro)
  e77bd2a  BLAKE3 in Dragline (commit-3, adr-fmt-xm8b)
  b9aad69  BLAKE3 follow-up (commit-3)

This is the 4th increment; sub-4b (Encode impls for gh-report DomainEvent)
and sub-4c (workspace verify) follow.

Report bead: adr-fmt-p9jw.

+++ crates/gh-report/src/domain/aggregates/run.rs
@@ -49,9 +49,13 @@ pub struct Run {
     /// `batch_id` from the originating `SweepStarted`, if any.
     pub batch_id: Option<String>,
     /// Total repo count declared at sweep start.
-    pub repo_count: usize,
+    //
+    // Width fixed at u64 per GEN-0004:R1 / GEN-0032 / GEN-0037 EVT-004:
+    // the `Run` aggregate's fields mirror the `DomainEvent` field widths,
+    // which are u64 to keep canonical-bytes target-independent.
+    pub repo_count: u64,
     /// Number of repos completed (last `SweepProgress::completed`).
-    pub completed: usize,
+    pub completed: u64,
 }
 
 impl Aggregate for Run {
@@ -131,7 +135,10 @@ pub enum RunError {
 #[derive(Debug, Clone)]
 pub struct StartSweep {
     pub org: String,
-    pub repo_count: usize,
+    // u64 mirrors the resulting `DomainEvent::SweepStarted.repo_count`
+    // (GEN-0004:R1 — no platform-dependent widths in domain types).
+    // Conversion from `usize` happens at the producer per COM-0023.
+    pub repo_count: u64,
     pub batch_id: String,
     pub timestamp: String,
 }
@@ -141,8 +148,8 @@ impl Command for StartSweep {}
 #[derive(Debug, Clone)]
 pub struct RecordProgress {
     pub batch_id: String,
-    pub completed: usize,
-    pub total: usize,
+    pub completed: u64,
+    pub total: u64,
     pub timestamp: String,
 }
 impl Command for RecordProgress {}
@@ -152,7 +159,7 @@ impl Command for RecordProgress {}
 pub struct CompleteSweep {
     pub batch_id: String,
     pub duration_ms: u64,
-    pub repo_count: usize,
+    pub repo_count: u64,
     pub timestamp: String,
 }
 impl Command for CompleteSweep {}
@@ -170,7 +177,7 @@ impl Command for FailSweep {}
 /// Publish evidence after a successful sweep.
 #[derive(Debug, Clone)]
 pub struct PublishEvidence {
-    pub page_count: usize,
+    pub page_count: u64,
     /// Whether this publish is a warm-start (replayed from baseline,
     /// no fresh GitHub API calls). Distinguishes cold-boot replay from
     /// live cycle output for downstream cache-warming consumers.
@@ -183,8 +190,8 @@ impl Command for PublishEvidence {}
 #[derive(Debug, Clone)]
 pub struct RenderPartial {
     pub batch_id: String,
-    pub page_count: usize,
-    pub pending_repos: usize,
+    pub page_count: u64,
+    pub pending_repos: u64,
     pub timestamp: String,
 }
 impl Command for RenderPartial {}
